db에서 데이터를 받아오는 dto는 internal dto로 넣음 + db에서 받아오는 것들에 대해서 추가
