--- spring-boot/src/main/java/com/demo/comentoStatistic/dto/LoginStatsDto.java
@@ -1,10 +1,0 @@
-package com.demo.comentoStatistic.dto;
-
-public record LoginStatsDto(
-        boolean isSuccess,
-        String yearMonth,
-        String totCnt,
-        double average,
-        String requestLog
-) {
-}
