+++ spring-boot/src/main/java/com/demo/comentoStatistic/dto/internal/LoginStatsDto.java
@@ -1,0 +1,8 @@
+package com.demo.comentoStatistic.dto.internal;
+
+public record LoginStatsDto(
+        String yearMonth,
+        Long totCnt,
+        String requestLog
+) {
+}
